refactor(nuker): extract wait functions to k8s pod
